Task with Mammals OOD is added. (not polymorph)
